style: move social link/resume under bio in main column
