+++ src/components/SideBar/MenuListItem.scss
@@ -4,10 +4,6 @@
   color: #fff !important;
 }
 
-.headline .headline-list-item {
-  padding: 8px 48px !important;
-}
-
 .headline .headline-text span {
   font-weight: bold !important;
 }
