+++ src/components/SideBar/SideBar.scss
@@ -1,7 +1,7 @@
 @import "../../Model.scss";
 
 .side-bar {
-  width: 20%;
+  width: 15%;
   float: left;
   color: $white;
   text-align: center;
