+++ src/components/TopPage/TopPage.tsx
@@ -1,58 +1,65 @@
 import React from "react";
 import "./TopPage.scss";
 import { Link as RouterLink } from "react-router-dom";
-import Link from "@material-ui/core/Link";
-import { Favorite, FindReplace, Group } from "@material-ui/icons";
+import { Link, Button } from "@material-ui/core";
+import { Favorite, FindReplace, Group, GitHub, LinkedIn, Twitter, Description } from "@material-ui/icons";
 import { portfolioData } from "../../data/portfolio";
+import IconButtonLink from "./IconButtonLink";
+
+const iconData = [
+  { url: "https://github.com/nk18chi", element: <GitHub /> },
+  { url: "https://www.linkedin.com/in/naoki-mita-ab186839/", element: <LinkedIn /> },
+  { url: "https://twitter.com/naokien1", element: <Twitter /> },
+  { url: "https://leetcode.com/nk18chi/", element: <img src='/portfolio/img/icon/leetcode.svg' alt='' /> },
+];
+const skillSets = [
+  {
+    title: "Typescript & React",
+    description:
+      "I have experience in building a React app with Firebase. I am familiar with React Hooks. I can program by Javascript but I prefer typescript to javascript. It helps me to make apps safe and find the bug easily.",
+    icons: ["/portfolio/img/icon/typescript.svg", "/portfolio/img/icon/react.svg"],
+  },
+  {
+    title: "Python & Django",
+    description:
+      "I have experience in publishing some apps that are developed by Django and MySQL. Also, I can make use of Rest API to handle data from the database and write test cases by fixtures and factory-boy.",
+    icons: ["/portfolio/img/icon/python.svg", "/portfolio/img/icon/django.svg"],
+  },
+  {
+    title: "Data Analysis",
+    description:
+      "I know how to get data from google analytics by Python(via API) or BigQuery and data from the database by SQL. I can analyze how users use your product by investigating those data and come up with a new strategy.",
+    icons: ["/portfolio/img/icon/google-analytics.svg", "/portfolio/img/icon/google-big-query.png", "/portfolio/img/icon/sql.svg"],
+  },
+  {
+    title: "Mobile Apps",
+    description:
+      "The experience in building web apps is bigger than one in mobile apps but I have knowledge of mobile apps as well and I am able to simple apps in iOS(Swift) and Android(Kotlin).",
+    icons: ["/portfolio/img/icon/apple.svg", "/portfolio/img/icon/android.svg"],
+  },
+];
+
+const myValues = [
+  {
+    icon: <Favorite />,
+    title: "Passion for building apps",
+    description: "I love building apps and thinking about how make users comfortable. I never stop learning and building the app even free time.",
+  },
+  {
+    icon: <FindReplace />,
+    title: "Think efficiently",
+    description: "I always keep looking for the way to reduce time. I love automation and effective-algorithm.",
+  },
+  {
+    icon: <Group />,
+    title: "Work together well",
+    description:
+      "I can smoothly work with a team from other departments because I have knowledge of not only building apps but also analysis and marketing, design etc.",
+  },
+];
 
 const TopPage: React.FC = () => {
   document.title = "Naoki Mita - portfolio, online resume";
-  const skillSets = [
-    {
-      title: "Typescript & React",
-      description:
-        "I have experience in building a React app with Firebase. I am familiar with React Hooks. I can program by Javascript but I prefer typescript to javascript. It helps me to make apps safe and find the bug easily.",
-      icons: ["/portfolio/img/icon/typescript.svg", "/portfolio/img/icon/react.svg"],
-    },
-    {
-      title: "Python & Django",
-      description:
-        "I have experience in publishing some apps that are developed by Django and MySQL. Also, I can make use of Rest API to handle data from the database and write test cases by fixtures and factory-boy.",
-      icons: ["/portfolio/img/icon/python.svg", "/portfolio/img/icon/django.svg"],
-    },
-    {
-      title: "Data Analysis",
-      description:
-        "I know how to get data from google analytics by Python(via API) or BigQuery and data from the database by SQL. I can analyze how users use your product by investigating those data and come up with a new strategy.",
-      icons: ["/portfolio/img/icon/google-analytics.svg", "/portfolio/img/icon/google-big-query.png", "/portfolio/img/icon/sql.svg"],
-    },
-    {
-      title: "Mobile Apps",
-      description:
-        "The experience in building web apps is bigger than one in mobile apps but I have knowledge of mobile apps as well and I am able to simple apps in iOS(Swift) and Android(Kotlin).",
-      icons: ["/portfolio/img/icon/apple.svg", "/portfolio/img/icon/android.svg"],
-    },
-  ];
-
-  const myValues = [
-    {
-      icon: <Favorite />,
-      title: "Passion for building apps",
-      description: "I love building apps and thinking about how make users comfortable. I never stop learning and building the app even free time.",
-    },
-    {
-      icon: <FindReplace />,
-      title: "Think efficiently",
-      description: "I always keep looking for the way to reduce time. I love automation and effective-algorithm.",
-    },
-    {
-      icon: <Group />,
-      title: "Work together well",
-      description:
-        "I can smoothly work with a team from other departments because I have knowledge of not only building apps but also analysis and marketing, design etc.",
-    },
-  ];
-
   return (
     <article className='main-column'>
       <section id='about-me' className='about-me block'>
@@ -63,6 +70,18 @@ const TopPage: React.FC = () => {
             I am a software developer specialized in frontend and backend development for web apps. Also, I have experience of being a data analyst. I
             believe I can help you to grow up your products in every aspect. Check out my portfolio and online resume.
           </p>
+          <div className='social-link'>
+            <ul>{iconData.length > 0 && iconData.map((i: any, index: number) => <IconButtonLink key={index} url={i.url} icon={i.element} />)}</ul>
+          </div>
+          <Button
+            href='https://docs.google.com/document/d/1GVcJ-RGKe2-6iPmtFZniVHD4ZNM1zV9tcE2bVAeZGpU/edit'
+            className='black-btn'
+            variant='contained'
+            size='large'
+            startIcon={<Description />}
+          >
+            View Resume
+          </Button>
         </div>
         <div className='summary-image'>
           <img className='main-thumnail' src='/portfolio/img/about_me.jpg' alt='' />
