+++ src/index.scss
@@ -176,7 +176,7 @@ h1 {
 }
 
 .main-column {
-  width: 78%;
+  width: 83%;
   float: right;
   background: #fff;
   min-height: 750px;
